More balanced beauty, work and so.
I think your mod will be more balanced this way. 
The beauty of your dirt shouldn't be more than the beauty of wooden floor, who cost wood. ( and also: it's dirt)
Also, the different type of dirt shouldn't have the same beauty.

The purpose of your mod is not to have beautiful base: it's to build path and road.